refactor to use concurrency task
use ember-concurrency for async calls where the component might be destroyed
before the response is returned.

This is pretty well explained here: http://ember-concurrency.com/docs/tutorial

In short:  The problem is that it's possible for our promise callback to run after the component has been
destroyed, and Ember (and React and many others) will complain if you try and,
well, call set() on a destroyed object.