Correctly use const instead of let in method body
The method declares a bunch of functions with the name `const`.  When
one of them is a `let` one is made to assume this variable has
properties which change values.  But wait, there's a trick.  This is
not the case here.  In the name of cleanup, we're changing this to a
const so it follows the expectations of the unknowing reader.