Inlining patternMatch because we can call it only once
We can call patternMatch only once.  It could make sense to extract
the regex itself into a global variable and document it, but for a
single call it seems to make sense to inline it as it is here.

The code is becoming a bit more complex, so we'll be adding comments
or extracting more logic from it.