Match matchedTerm name with name of matched index
The matched term is now also the name of the thing we pass through.
This consistency brings value in that users expect to see the same
variable name if it is used in a similar context.  This is the case
for us, so if we can match up the variables whilst still having a
clear local context, we should.  This seems to hold here.