Introduce single solution answer and clean up code
By introducing a single solution, the code becomes cleaner and it
becomes easier to insert the necessory elements.  This is not to say
that the new code is ideal.  The code with the array was easier to
extend.

This new piece of code helps to ensure that we do not insert a link
when no article was found.  Nice.