Updated yuidoc strings
Some have been added.  I'm sure we don't extend Ember.Component
anymore, I'm not sure Glimmer.Component is now the right way of
mentioning this.  In fact, I'm not sure at al how we should refer to
such components.

This commit is an improvement, but I'm not sure it is right.  If
someone knows, I'd like to know. :)